Feature/auth admin (#25)
* Added Admin Feature

* Fixed Admin Delete

* Added more tests and some minor status code fixes

* Cleaned Up 404

* Changed Login Response

* Removed unused imports

+++ backend/services/user.ts
@@ -1,15 +1,27 @@
 import * as bcrypt from 'bcrypt';
 import { User, UserPublic } from "../models/user";
 import * as userDAO from "../storage/user";
+import * as statusDAO from "../storage/status";
 import { userTransformer } from '../util/transformer';
 
 export function createUser(user: User): Promise<UserPublic> {
+    //Default Role
+    user.role = "user";
+
+    //Create User
     return new Promise((resolve, reject) => {
         //BCrypt User´s Password
         bcrypt.hash(user.password, 10).then(hash => {
             user.password = hash;
-            //Save User in DB
-            userDAO.createUser(user).then(user => resolve(userTransformer(user))).catch(() => reject(new Error("User already exists")));
+            //Check for First User to create Admin
+            statusDAO.getCounterTable("user").then(counter => {
+                if (counter === 0) {
+                    user.role = "admin";
+                }
+            }).finally(() => {
+                //Save User in DB
+                userDAO.createUser(user).then(user => resolve(userTransformer(user))).catch(() => reject(new Error("User already exists")));
+            });
         });
     });
 };
@@ -54,8 +66,12 @@ export function deleteUser(id: string): Promise<boolean> {
         //Check if User exists
         userDAO.getUserById(id).then(user => {
             if (user) {
-                //Delete User from DB
-                userDAO.deleteUser(id).then(() => resolve(true)).catch(() => reject(new Error("Error Deleting User")));
+                if (user.role === "admin") {
+                    reject(new Error("Admin can´t be deleted"));
+                } else {
+                    //Delete User from DB
+                    userDAO.deleteUser(id).then(() => resolve(true)).catch(() => reject(new Error("Error Deleting User")));
+                }
             } else {
                 reject(new Error("User not found"));
             }
